% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_subcategories.R
\name{classify_subcategories}
\alias{classify_subcategories}
\title{Classifier for subcategory data, using nearest neighbours and master file data}
\usage{
classify_subcategories(df_new_expenses, path_category_dict, df_old_master_file)
}
\arguments{
\item{df_new_expenses}{Dataframe of new expenses to be categorised.}

\item{path_category_dict}{Path to dictionary.}

\item{df_old_master_file}{Dataframe of old master master file.}
}
\value{
Categorised master file.
}
\description{
Classifier for subcategory data, using nearest neighbours and master file data
}
