% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_converter.R
\name{convert_amount}
\alias{convert_amount}
\title{Convert currency using priceR package (and API)}
\usage{
convert_amount(df, currency_in, currency_out)
}
\arguments{
\item{df}{Dataframe with multiple currency columns, one of which (currency_in)
is to be converted to the other (curency_out).}

\item{currency_in}{Currency to be converted from.}

\item{currency_out}{Currency to be converted into.}
}
\value{
Converted dataframe.
}
\description{
Convert currency using priceR package (and API)
}
