% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{get_rates}
\alias{get_rates}
\title{Get rates using priceR}
\usage{
get_rates(currency_in, currency_out, date_start, date_end)
}
\arguments{
\item{date_end}{}
}
\value{
A data.frame with date and exchange columns.
}
\description{
Note: Requires internet connection to download currency conversions.
}
\examples{
get_rates("eur", "gbp", "2022-01-01", "2022-02-01")
}
