% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_boa.R
\name{import_boa}
\alias{import_boa}
\title{Process Bank of America file}
\usage{
import_boa(path, currency = "usd")
}
\arguments{
\item{path}{Path of raw Bank of America file}

\item{currency}{Currency of Bank of America file (USD)}
}
\value{
Same data in common data format.
}
\description{
Process Bank of America file
}
