% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_boa_credit.R
\name{import_boa_credit}
\alias{import_boa_credit}
\title{Process Bank of America credit card file. User must manually modify
statements to ensure correct date ranges}
\usage{
import_boa_credit(path, currency = "usd")
}
\arguments{
\item{path}{Path of raw Bank of America credit card file}

\item{currency}{Currency of Bank of America file (USD)}
}
\value{
Same data in common data format.
}
\description{
Process Bank of America credit card file. User must manually modify
statements to ensure correct date ranges
}
