% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_capitalone_debit.R
\name{import_capitalone_debit}
\alias{import_capitalone_debit}
\title{Process Capital One debit card file. User must manually modify
statements to ensure correct date ranges}
\usage{
import_capitalone_debit(path, currency = "usd")
}
\arguments{
\item{path}{Path of raw Capital One file}

\item{currency}{Currency of Capital One file (USD)}
}
\value{
Same data in common data format.
}
\description{
Process Capital One debit card file. User must manually modify
statements to ensure correct date ranges
}
