% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chase.R
\name{import_chase}
\alias{import_chase}
\title{Process Chase credit card file. User must manually modify
statements to ensure correct date ranges}
\usage{
import_chase(path, currency = "usd")
}
\arguments{
\item{path}{Path of raw Chase credit card file}

\item{currency}{Currency of Chase file (USD)}
}
\value{
Same data in common data format.
}
\description{
Process Chase credit card file. User must manually modify
statements to ensure correct date ranges
}
