% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_danske.R
\name{import_danske}
\alias{import_danske}
\title{Process Danske Bank file}
\usage{
import_danske(path, currency = "dkk")
}
\arguments{
\item{path}{Path of raw Danske Bank file.}

\item{currency}{Currency of Danske Bank file (DKK).}
}
\value{
Same data in common data format.
}
\description{
Process Danske Bank file
}
