% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_natwest.R
\name{import_natwest}
\alias{import_natwest}
\title{Process Natwest file}
\usage{
import_natwest(path, currency = "gbp")
}
\arguments{
\item{path}{Path of raw Natwest file}

\item{currency}{Currency of Natwest file (GBP).}
}
\value{
Same data in common data format.
}
\description{
Process Natwest file
}
