% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_revolut.R
\name{import_revolut}
\alias{import_revolut}
\title{Process Revolut file}
\usage{
import_revolut(path, currency)
}
\arguments{
\item{path}{Path of raw Revolut file}

\item{currency}{Currency of Revolut file}
}
\value{
Same data in common data format.
}
\description{
Process Revolut file
}
