% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_santander.R
\name{import_santander}
\alias{import_santander}
\title{Process Santander file}
\usage{
import_santander(path, currency = "usd")
}
\arguments{
\item{path}{Path of raw Santander file}

\item{currency}{Currency of Santander file (USD)}
}
\value{
Same data in common data format.
}
\description{
Process Santander file
}
