% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balances.R
\name{plot_balances}
\alias{plot_balances}
\title{Create balances plot}
\usage{
plot_balances(df)
}
\arguments{
\item{df_balances}{Dataframe containing raw balances data}

\item{currency}{User inputted currency for balances plot}
}
\value{
Plotly graph
}
\description{
Create balances plot
}
