% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Create waterfall plot}
\usage{
plot_waterfall(df)
}
\arguments{
\item{df}{Dataframe containing processed master file data}
}
\value{
Plotly graph
}
\description{
Create waterfall plot
}
