% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf-8_encoder.R
\name{to_utf8}
\alias{to_utf8}
\title{Convert path to file to UTF-8 encoding to avoid character issues}
\usage{
to_utf8(path)
}
\arguments{
\item{path}{The path to a file.}
}
\value{
The UTF-8 encoded input text (invisibly).
}
\description{
Convert path to file to UTF-8 encoding to avoid character issues
}
