% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{transform_master_for_waterfall}
\alias{transform_master_for_waterfall}
\title{Manipulate master file data in preparation for waterfall plot}
\usage{
transform_master_for_waterfall(df, master_currency_analytics)
}
\arguments{
\item{df}{Dataframe containing master file data}

\item{master_currency_analytics}{User inputted master currency for waterfall plot}
}
\value{
Dataframe to be passed to plot_waterfall() function
}
\description{
Manipulate master file data in preparation for waterfall plot
}
