% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{classify_subcategories}
\alias{classify_subcategories}
\title{Nearest neighbour subcategory classifier using Jaccard similarity on shingles}
\usage{
classify_subcategories(df_new_expenses, path_category_dict, latest_master_file)
}
\arguments{
\item{latest_master_file}{}
}
\description{
Nearest neighbour subcategory classifier using Jaccard similarity on shingles
}
