% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{convert_amount}
\alias{convert_amount}
\title{Convert expenses using exchanges data}
\usage{
convert_amount(df, df_exchanges)
}
\arguments{
\item{df_exchanges}{}
}
\value{
A data.frame with new columns for converted amounts.
}
\description{
Convert expenses using exchanges data
}
